pyproject.toml
[project]
name = "newstrendssummary"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
    "schedule>=1.2.0",
    "scikit-learn>=1.8.0",
]
